+++ nflow-engine/src/test/java/io/nflow/engine/internal/dao/StatisticsDaoTest.java
@@ -11,6 +11,7 @@ import static org.hamcrest.Matchers.nullValue;
 import static org.joda.time.DateTime.now;
 
 import java.util.Map;
+import java.util.Optional;
 
 import jakarta.inject.Inject;
 
@@ -74,7 +75,7 @@ public class StatisticsDaoTest extends BaseDaoTest {
     assertThat(stateStats.get(created.name()).allInstances, is(1L));
 
     WorkflowInstance i2 = new WorkflowInstance.Builder().setId(id).setNextActivation(now().minusDays(1)).build();
-    instanceDao.updateNotRunningWorkflowInstance(i2);
+    instanceDao.updateNotRunningWorkflowInstance(i2, Optional.empty());
 
     stats = statisticsDao.getWorkflowDefinitionStatistics(i1.type, null, null, null, null);
     stateStats = stats.get(TestState.BEGIN.name());